Remove service-token rotation; address PR 421 review
Apply the non-Codex review fixes from PR 421:
- Clamp cron widget toggleDay to keep >=1 day in "week" mode
- Simplify CronWidgetMap type and drop unused RegistryWidgetsType import
- Mark ChatKit domain key optional in stack .env.example
- Drop redundant ^/$ anchors from the configurable-template regex

Remove the service-token rotate action end-to-end (Studio button, the
`orcheo token rotate` CLI command, the SDK service helper, the backend
/rotate endpoint and schema, and ServiceTokenManager.rotate), prune the
corresponding tests, and update the CLI/auth/ChatKit docs. The
storage-level rotated_to/rotation_expires_at fields and "Rotated" status
display are kept so any pre-existing rotated tokens still render.
