+++ apps/studio/src/lib/api.ts
@@ -505,22 +505,6 @@ export async function createServiceToken(
   );
 }
 
-export async function rotateServiceToken(
-  tokenId: string,
-  overlapSeconds: number,
-  baseUrl?: string,
-): Promise<ServiceToken> {
-  return serviceTokenRequest<ServiceToken>(
-    `/api/admin/service-tokens/${encodeURIComponent(tokenId)}/rotate`,
-    {
-      method: "POST",
-      body: JSON.stringify({ overlap_seconds: overlapSeconds }),
-    },
-    "Failed to rotate service token",
-    baseUrl,
-  );
-}
-
 export async function revokeServiceToken(
   tokenId: string,
   reason: string,
