+++ deploy/stack/.env.example
@@ -17,8 +17,10 @@ ORCHEO_VAULT_ENCRYPTION_KEY=replace-with-64-hex-chars
 ORCHEO_CHATKIT_TOKEN_SIGNING_KEY=strong-random-secret
 # Bootstrap API token used when auth mode is `required` or `optional`.
 ORCHEO_AUTH_BOOTSTRAP_SERVICE_TOKEN=change-me
-# Public ChatKit domain key consumed by Studio; required for ChatKit UI features.
-# Security note: this value is shipped to browsers. Scope it per environment.
+# Public ChatKit domain key consumed by Studio; optional — set it to your
+# ChatKit domain key to enable ChatKit UI features (a placeholder is used when
+# left unset). Security note: this value is shipped to browsers. Scope it per
+# environment.
 VITE_ORCHEO_CHATKIT_DOMAIN_KEY=domain_pk_replace_me
 # HMAC secret used by the Credential Broker to sign run-scoped tokens.
 # Auto-generated by `orcheo install` on a fresh stack env file.
